refactor(chat): collapse duplicate ApprovalModal/ApprovalOverlay into shared ApprovalDialog
The corpus-chat ApprovalModal and document-tray ApprovalOverlay were ~90%
identical copy-paste — same markup, same flex-column scroll structure, only
cosmetic icon/prop-name divergence. Replaced with one shared
frontend/src/components/chat/ApprovalDialog.tsx. Both CorpusChat and
ChatTray now render it directly.

Standardized icons (AlertTriangle header, XCircle reject, lucide <X> close
with aria-label). Removed redundant wordBreak: break-word (superseded by
overflowWrap: anywhere). PendingApproval type re-exports repointed from
./ApprovalOverlay to the canonical components/chat/types in the three
handler modules and their tests.

Renamed tests/ApprovalModal.ct.tsx -> tests/ApprovalDialog.ct.tsx and the
screenshot tag chat--approval-modal--pending -> chat--approval-dialog--pending.

Addresses the DRY note from the Claude review on PR #1787; the issue-#2
redundant-CSS-prop nit is also folded in. Issue-#1 (viewport breathing-room
on maxHeight: 100%) intentionally deferred — out of scope for this DRY pass.

All 7 CT tests + 90 hook unit tests pass.
