+++ CHANGELOG.md
@@ -9,7 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- **Tool approval modal overflows the viewport when arguments are large** (`frontend/src/components/corpuses/corpus_chat/ApprovalModal.tsx`, `frontend/src/components/knowledge_base/document/right_tray/ApprovalOverlay.tsx`). Both approval popups for pending agent tool calls rendered the tool's JSON `arguments` inside a `<pre style={{ whiteSpace: "pre-wrap" }}>` nested in an inner modal that had only `maxWidth: 500px` — no `maxHeight`, no scroll container. When a tool call carried a multi-thousand-character payload (e.g. `update_corpus_description` with a long `new_content` string), the inner modal grew unbounded, the Approve / Reject buttons were pushed off-screen, and the user couldn't act on the request. Fix: inner modal is now a `display: flex; flex-direction: column` box capped at `maxHeight: 100%` with `boxSizing: border-box`; the body section (`flex: 1; minHeight: 0; overflowY: auto`) scrolls; header and footer (`flexShrink: 0`) stay pinned so the action buttons remain visible. The `<pre>` gains `wordBreak: break-word` + `overflowWrap: anywhere` so a single unbroken token can't blow out the width. New regression test `ApprovalModal › constrains modal height and scrolls body when args are very long` (`frontend/tests/ApprovalModal.ct.tsx`) mounts the modal with ~10k-char `new_content` arguments and asserts both buttons stay within the viewport's `boundingBox`.
+- **Tool approval modal overflows the viewport when arguments are large** (`frontend/src/components/chat/ApprovalDialog.tsx`, `frontend/src/components/corpuses/CorpusChat.tsx`, `frontend/src/components/knowledge_base/document/right_tray/ChatTray.tsx`). Both approval popups for pending agent tool calls rendered the tool's JSON `arguments` inside a `<pre style={{ whiteSpace: "pre-wrap" }}>` nested in an inner modal that had only `maxWidth: 500px` — no `maxHeight`, no scroll container. When a tool call carried a multi-thousand-character payload (e.g. `update_corpus_description` with a long `new_content` string), the inner modal grew unbounded, the Approve / Reject buttons were pushed off-screen, and the user couldn't act on the request. Fix: inner modal is now a `display: flex; flex-direction: column` box capped at `maxHeight: 100%` with `boxSizing: border-box`; the body section (`flex: 1; minHeight: 0; overflowY: auto`) scrolls; header and footer (`flexShrink: 0`) stay pinned so the action buttons remain visible. The `<pre>` gains `overflowWrap: anywhere` so a single unbroken token can't blow out the width. New regression test `ApprovalDialog › constrains modal height and scrolls body when args are very long` (`frontend/tests/ApprovalDialog.ct.tsx`) mounts the modal with ~10k-char `new_content` arguments and asserts both buttons stay within the viewport's `boundingBox`.
+
+### Changed
+
+- **Approval modal DRY pass: collapsed the two ~90%-duplicated approval popups into a single `ApprovalDialog`** (`frontend/src/components/chat/ApprovalDialog.tsx`). The corpus-chat `ApprovalModal` (`frontend/src/components/corpuses/corpus_chat/ApprovalModal.tsx`) and the document-tray `ApprovalOverlay` (`frontend/src/components/knowledge_base/document/right_tray/ApprovalOverlay.tsx`) were near-identical copy-paste — same markup, same flex column scroll structure, with only cosmetic divergence (header icon `AlertCircle` vs `AlertTriangle`, Reject icon `X` vs `XCircle`, unicode `✕` vs lucide `<X>` for the close button, slightly different prop naming). They have been replaced with one shared `ApprovalDialog` component in `frontend/src/components/chat/ApprovalDialog.tsx` (the chat-domain sibling of the existing `RequestingAgentAttribution`). `CorpusChat` (`frontend/src/components/corpuses/CorpusChat.tsx`) and `ChatTray` (`frontend/src/components/knowledge_base/document/right_tray/ChatTray.tsx`) now render `ApprovalDialog` directly with `{ pendingApproval, show, onHide, onDecision }` props. The old `ApprovalModal.tsx` file is deleted; `ApprovalOverlay.tsx` is reduced to just the small `ReopenApprovalButton` that's only relevant to the document tray. Standardized icons (AlertTriangle for the header, XCircle for Reject, lucide `<X>` for close with an `aria-label`) and removed the redundant `wordBreak: break-word` (superseded by `overflowWrap: anywhere`) per Claude review on PR #1787. `PendingApproval` re-export pointers in `useChatSendHandlers` / `useChatStreamHandlers` / `useChatAgentMessageHandler` and their unit tests repoint from `./ApprovalOverlay` to the canonical `components/chat/types`. The CT test file is renamed `frontend/tests/ApprovalModal.ct.tsx` → `frontend/tests/ApprovalDialog.ct.tsx` (and screenshot tag `chat--approval-modal--pending` → `chat--approval-dialog--pending`). All 7 CT tests + 90 hook unit tests pass.
 
 ### Changed
